+++ src/components/manage/BlockChooser/BlockChooser.test.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { map } from 'lodash';
 import '@testing-library/jest-dom/extend-expect';
 import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-intl-redux';
@@ -238,9 +239,11 @@ describe('BlocksChooser', () => {
     expect(container).not.toHaveTextContent('Custom group');
     config.blocks.groupBlocksOrder = old;
   });
-  it('Show unique in case they have been removed', () => {
+  it('Show unique in case they have been removed (restricted as a function)', () => {
     config.blocks.required = [];
-    config.blocks.blocksConfig.title.unique = true;
+    config.blocks.blocksConfig.title.restricted = ({ properties, block }) =>
+      map(properties.blocks, (item) => item['@type']).includes(block.id);
+
     const { container } = render(
       <Provider store={store}>
         <BlockChooser
@@ -260,7 +263,8 @@ describe('BlocksChooser', () => {
   });
   it('Do not show unique in case it is present', () => {
     config.blocks.required = [];
-    config.blocks.blocksConfig.title.unique = true;
+    config.blocks.blocksConfig.title.restricted = true;
+
     const { container } = render(
       <Provider store={store}>
         <BlockChooser